+++ src/components/CurrentLessonInfo/CurrentLessonInfo.styled.ts
@@ -1,0 +1,15 @@
+import styled from "styled-components";
+
+export const CurrentLessContainer = styled.div`
+  margin-bottom: 20px;
+`;
+
+export const CurrentLesson = styled.p`
+  margin: 8px 0;
+  font-size: 20px;
+  font-weight: 600;
+`;
+
+export const CurrentText = styled.span`
+  font-weight: 400;
+`;
