+++ src/components/LessonsList/LessonsList.tsx
@@ -1,17 +1,8 @@
 import React from "react";
-import {
-  CourseLesson,
-  CourseLessonsList,
-  CurrentLessContainer,
-  CurrentLesson,
-  CurrentText,
-  LessonTitle,
-  LockIcon,
-  Title,
-  TitleBox,
-} from "./LessonsList.styled";
+import { CourseLessonsList, Title } from "./LessonsList.styled";
 import { ILesson, LessonseDetails } from "types/type";
-import { LOCKED } from "constants/other";
+import { CurrentLessonInfo } from "components/CurrentLessonInfo/CurrentLessonInfo";
+import { LessonItem } from "components/LessonItem/LessonItem";
 
 interface ILessonsListProps {
   lessons?: ILesson[];
@@ -27,30 +18,20 @@ const LessonsList: React.FC<ILessonsListProps> = ({
   return (
     <>
       {currentLesson && (
-        <CurrentLessContainer>
-          <CurrentLesson>
-            Current Lesson: <CurrentText>{currentLesson.title}</CurrentText>
-          </CurrentLesson>
-          <CurrentLesson>
-            Status: <CurrentText>{currentLesson.status}</CurrentText>
-          </CurrentLesson>
-        </CurrentLessContainer>
+        <CurrentLessonInfo
+          title={currentLesson?.title}
+          status={currentLesson?.status}
+        />
       )}
       <Title>Lessons:</Title>
       <CourseLessonsList>
         {lessons?.map((lesson: LessonseDetails) => (
-          <CourseLesson
-            onClick={() => onLessonClick(lesson)}
+          <LessonItem
+            handleLessonClick={() => onLessonClick(lesson)}
             key={lesson.id}
-            id={lesson.id}
-          >
-            <TitleBox>
-              <LessonTitle active={currentLesson?.id === lesson.id}>
-                {lesson.title}.
-              </LessonTitle>
-              {lesson.status === LOCKED && <LockIcon />}
-            </TitleBox>
-          </CourseLesson>
+            lesson={lesson}
+            currentLessonId={currentLesson?.id}
+          />
         ))}
       </CourseLessonsList>
     </>
