+++ src/redux/auth/authApi.ts
@@ -1,5 +1,4 @@
 import { createApi, fetchBaseQuery } from "@reduxjs/toolkit/query/react";
-import { REHYDRATE } from "redux-persist";
 
 interface AuthResponse {
   token: string;
@@ -18,11 +17,6 @@ export const authApi = createApi({
   }),
   reducerPath: "authApi",
   tagTypes: ["Auth"],
-  extractRehydrationInfo(action, { reducerPath }) {
-    if (action.type === REHYDRATE) {
-      return action.payload[reducerPath];
-    }
-  },
   endpoints: (builder) => ({
     getToken: builder.query<AuthResponse, string>({
       query: () => "",
